l9110/host/requirements.txt
numpy
scipy
pygame-ce
